+++ packages/gator-permissions-controller/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.3.0]
 
+### Added
+
+- Add `submitRevocation` action to submit permission revocations through the gator permissions provider snap
+- Add `addPendingRevocation` action to queue revocations until transaction confirmation
+
 ### Changed
 
 - **BREAKING:** Use new `Messenger` from `@metamask/messenger` ([#6461](https://github.com/MetaMask/core/pull/6461))